Extracts connectionlist class
